Remove limit from SelectInput
