+++ packages/cli-kit/src/private/node/ui/components/SelectInput.tsx
@@ -1,27 +1,12 @@
 import React, {useState, useEffect, useRef, useCallback} from 'react'
-import {Box, Text, useInput} from 'ink'
+import {Box, Text, useApp, useInput} from 'ink'
 import {isEqual} from 'lodash-es'
-
-function arrayRotate<T>(array: T[], count?: number) {
-  const slice = array.slice()
-  return slice.splice(-(count ?? 0) % slice.length).concat(slice)
-}
 
 export interface Props {
   /**
    * Items to display in a list. Each item must be an object and have `label` and `value` props.
    */
   items?: Item[]
-
-  /**
-   * Index of initially-selected item in `items` array.
-   */
-  initialIndex?: number
-
-  /**
-   * Number of items to display.
-   */
-  limit?: number
 
   /**
    * Function to call when user selects an item. Item object is passed to that function as an argument.
@@ -35,12 +20,10 @@ export interface Item {
   key?: string
 }
 
-const SelectInput: React.FC<Props> = ({items = [], initialIndex = 0, limit: customLimit, onSelect}): JSX.Element => {
-  const [rotateIndex, setRotateIndex] = useState(0)
-  const [selectedIndex, setSelectedIndex] = useState(initialIndex)
-  const hasLimit = typeof customLimit === 'number' && items.length > customLimit
-  const limit = hasLimit ? Math.min(customLimit, items.length) : items.length
+const SelectInput: React.FC<Props> = ({items = [], onSelect}): JSX.Element => {
+  const [selectedIndex, setSelectedIndex] = useState(0)
   const keys = useRef(new Set(items.map((item) => item.key)))
+  const {exit: unmountInk} = useApp()
 
   const previousItems = useRef<Item[]>(items)
 
@@ -51,7 +34,6 @@ const SelectInput: React.FC<Props> = ({items = [], initialIndex = 0, limit: cust
         items.map((item) => item.value),
       )
     ) {
-      setRotateIndex(0)
       setSelectedIndex(0)
     }
 
@@ -73,37 +55,28 @@ const SelectInput: React.FC<Props> = ({items = [], initialIndex = 0, limit: cust
         }
 
         if (key.upArrow) {
-          const lastIndex = (hasLimit ? limit : items.length) - 1
+          const lastIndex = items.length - 1
           const atFirstIndex = selectedIndex === 0
-          const nextIndex = hasLimit ? selectedIndex : lastIndex
-          const nextRotateIndex = atFirstIndex ? rotateIndex + 1 : rotateIndex
-          const nextSelectedIndex = atFirstIndex ? nextIndex : selectedIndex - 1
+          const nextSelectedIndex = atFirstIndex ? lastIndex : selectedIndex - 1
 
-          setRotateIndex(nextRotateIndex)
           setSelectedIndex(nextSelectedIndex)
         } else if (key.downArrow) {
-          const atLastIndex = selectedIndex === (hasLimit ? limit : items.length) - 1
-          const nextIndex = hasLimit ? selectedIndex : 0
-          const nextRotateIndex = atLastIndex ? rotateIndex - 1 : rotateIndex
-          const nextSelectedIndex = atLastIndex ? nextIndex : selectedIndex + 1
+          const atLastIndex = selectedIndex === items.length - 1
+          const nextSelectedIndex = atLastIndex ? 0 : selectedIndex + 1
 
-          setRotateIndex(nextRotateIndex)
           setSelectedIndex(nextSelectedIndex)
         } else if (key.return) {
-          const slicedItems = hasLimit ? arrayRotate(items, rotateIndex).slice(0, limit) : items
-
-          onSelect(slicedItems[selectedIndex]!)
+          unmountInk()
+          onSelect(items[selectedIndex]!)
         }
       },
-      [hasLimit, limit, rotateIndex, selectedIndex, items, onSelect],
+      [selectedIndex, items, onSelect],
     ),
   )
 
-  const slicedItems = hasLimit ? arrayRotate(items, rotateIndex).slice(0, limit) : items
-
   return (
     <Box flexDirection="column">
-      {slicedItems.map((item, index) => {
+      {items.map((item, index) => {
         const isSelected = index === selectedIndex
 
         return (
